[Add #234] GateStartButton.tsx 생성
